hardware/lge: add a hack to get DAC working with OSS audio HAL
This is an awful implementation to enable the Hi-Fi Quad DAC feature.

Due to cloning our own audio HAL causing several Soong issues, instead
opt to hijack the audio_amplifier HAL.

This has the limitation that we cannot introduce our own snd_device,
backend, or interface directly to the platform list in the HAL; to
address this, we can force SND_DEVICE_OUT_HEADPHONES to switch to LG's
custom backend and use those in place of the default backend. When
disabling, the original backend should be brought back.

To use this, define TARGET_DAC_PLUGIN_INCLUDE_DIR to your device
repository's header path, copy dac_plugin.h to that directory, and
rename the mixer paths, backends and interfaces as needed.

Change-Id: I4d61a288746c2eb34849831d58633c298e04622e