dac_panel: instantiate the QuadDAC class object separately
The DAC panel, both panel and tile, is not expected to stay alive in
the background.

This causes issues with the tile; it expects the QuadDAC class to
already be initialized, but this only happens in the panel, so without
having opened the panel, QuadDAC class may be uninitialized, marking
the tile as "unavailable".

To fix this, make the panel and tile use its own instances of the
QuadDAC class.

Change-Id: I6680b45f7cdc90261c932409494c667c5fd1d5b4